pyproject.toml
[tool.poetry]
name = "sck-core-codecommit"
version = "0.0.7"
description = "Core automstion codecommit listener"
authors = ["James Barwick <jbarwick@me.com>"]
license = "MIT"
readme = "README.md"
maintainers = [ "James Barwick <jbarwick@me.com>" ]
include = ["core_codecommit/version.py"]
keywords = ["sck", "core", "aws", "cloud", "automation"]

[[tool.poetry.packages]]
include = "core_codecommit"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
urllib3 = "^2.2.3"
boto3 = ">=1.39.9"
sck-core-framework = "^0.0.7"
sck-core-api = "^0.0.7"

[[tool.poetry.source]]
name = "nexus"
url = "https://monster-jj.jvj28.com:9091/repository/pypi/simple/"
priority = "primary"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
flake8 = "^7.1.1"
cfn-lint = "^1.22.3"
black = "^24.10.0"
pytest-dotenv = "^0.5.2"
boto3-stubs = ">=1.39.9"
pytest-cov = "^6.0.0"

[tool.poetry-dynamic-versioning]
enable = true
bump = true
ignore-untracked = true
style = "semver"

[tool.poetry-dynamic-versioning.substitution]
files = ["core_codecommit/__init__.py"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

